+++ src/components/Pricing/PricingExperiment.tsx
@@ -12,7 +12,7 @@ import Lottie from 'react-lottie'
 import Plans, { CTA as PlanCTA, PricingTiers } from './Plans'
 import Link from 'components/Link'
 import CTA from 'components/Home/CTA.js'
-import { IconCheck, IconGraph, IconRewindPlay, IconToggle, IconFlask, IconMessage, IconHandMoney, IconInfo, IconRocket, IconStarFilled, IconStar, IconMinus, IconPlus, IconChevronDown } from '@posthog/icons'
+import { IconCheck, IconGraph, IconRewindPlay, IconToggle, IconFlask, IconMessage, IconHandMoney, IconInfo, IconRocket, IconStarFilled, IconStar } from '@posthog/icons'
 import * as Icons from '@posthog/icons'
 import Tooltip from 'components/Tooltip'
 import useProducts from './Products'
@@ -23,294 +23,9 @@ import { CallToAction } from 'components/CallToAction'
 import Tabbed from './PricingCalculator/Tabbed'
 import { usePlatform } from './Platform/usePlatform'
 import { motion } from 'framer-motion'
-
-const PlanColumns = ({ billingProducts, highlight = 'paid' }) => {
-    const platformAndSupportProduct = billingProducts.find(
-        (product: BillingProductV2Type) => product.type === 'platform_and_support'
-    )
-    const highestSupportPlan = platformAndSupportProduct?.plans?.slice(-1)[0]
-
-    const [isPlanComparisonVisible, setIsPlanComparisonVisible] = useState(false)
-    return (
-        <>
-            <section className={`${section} mb-12 mt-8 md:px-4`}>
-                <h3 className="border-b border-light dark:border-dark pb-2 mb-6">Platform plans</h3>
-                <p className="text-[15px] text-primary/75 dark:text-primary-dark/75">
-                    All plans include unlimited team members and no limits on tracked users.
-                </p>
-                <div className="col-span-4 -mx-4 lg:mx-0 mb-4 px-4 lg:px-0 overflow-x-auto">
-                    <div
-                        className={`grid grid-cols-[repeat(3,_minmax(260px,_1fr))] xl:max-w-4xl xl:mx-auto gap-4 mb-12 ${highlight === 'free'
-                            ? '[&>*:nth-child(1)_>div]:border-red [&>*:nth-child(1)_>div]:border-3'
-                            : '[&>*:nth-child(2)_>div]:border-red [&>*:nth-child(2)_>div]:border-3'
-                            }`}
-                    >
-                        {planSummary.map((plan, index) => (
-                            <Plan key={index} planData={plan} highlight={plan.intent === highlight} />
-                        ))}
-                    </div>
-                </div>
-                <p
-                    className="text-center text-red dark:text-yellow font-bold cursor-pointer flex items-center justify-center"
-                    onClick={() => setIsPlanComparisonVisible(!isPlanComparisonVisible)}
-                >
-                    {isPlanComparisonVisible ? (
-                        <>
-                            Hide full plan comparison <IconChevronDown className="w-8 rotate-180" />
-                        </>
-                    ) : (
-                        <>
-                            Show full plan comparison <IconChevronDown className="w-8" />
-                        </>
-                    )}
-                </p>
-            </section>
-
-            <section
-                className={`${section} ${isPlanComparisonVisible
-                    ? 'visible max-h-full opacity-1 mb-12 mt-8 md:px-4'
-                    : 'overflow-y-hidden invisible max-h-0 opacity-0'
-                    } transition duration-500 ease-in-out transform`}
-            >
-                <div className="overflow-x-auto -mx-4 px-4 md:mx-0 md:px-0">
-                    <div className="grid grid-cols-16 mb-1 min-w-[1000px]">
-                        <div className="col-span-4 px-3 py-1">&nbsp;</div>
-                        {platformAndSupportProduct?.plans
-                            ?.filter((plan: BillingV2PlanType) => plan.name !== 'Teams') // This is a temporary addition until the teams addon is shipped and the teams plan is removed
-                            ?.map((plan: BillingV2PlanType) => (
-                                <div className="col-span-4 px-3 py-1" key={plan.key}>
-                                    <strong className="text-sm opacity-75">{plan.name}</strong>
-                                </div>
-                            ))}
-                    </div>
-
-                    <div className="grid grid-cols-16 mb-2 border-x border-b border-light dark:border-dark bg-white dark:bg-accent-dark [&>div]:border-t [&>div]:border-light dark:[&>div]:border-dark min-w-[1000px]">
-                        <div className="col-span-4 bg-accent/50 dark:bg-black/75 px-3 py-2 text-sm">
-                            <strong className="text-primary/75 dark:text-primary-dark/75">Base price</strong>
-                        </div>
-                        {platformAndSupportProduct?.plans
-                            ?.filter((plan: BillingV2PlanType) => plan.name !== 'Teams') // This is a temporary addition until the teams addon is shipped and the teams plan is removed
-                            ?.map((plan: BillingV2PlanType) => {
-                                return (
-                                    <div className="col-span-4 px-3 py-2 text-sm" key={`${plan.key}-base-price`}>
-                                        {plan.included_if === 'no_active_subscription' ? (
-                                            <span>Free forever</span>
-                                        ) : plan.included_if === 'has_subscription' ? (
-                                            <span>$0</span>
-                                        ) : plan.unit_amount_usd ? (
-                                            `$${parseFloat(plan.unit_amount_usd).toFixed(0)}/mo`
-                                        ) : plan.contact_support ? (
-                                            'Contact us'
-                                        ) : (
-                                            'Contact us'
-                                        )}
-                                    </div>
-                                )
-                            })}
-                        {highestSupportPlan?.features
-                            ?.filter(
-                                (f: BillingV2FeatureType) =>
-                                    ![
-                                        // TODO: this shouldn't be necessary, update billing products api to include entitlement_only info
-                                        'role_based_access',
-                                        'project_based_permissioning',
-                                        'ingestion_taxonomy',
-                                        'tagging',
-                                    ].includes(f.key)
-                            )
-                            .map((feature: BillingV2FeatureType) => (
-                                <>
-                                    <div className="col-span-4 bg-accent/50 dark:bg-black/75 px-3 py-2 text-sm">
-                                        {feature.description ? (
-                                            <Tooltip content={feature.description}>
-                                                <strong className="border-b border-dashed border-light dark:border-dark cursor-help text-primary/75 dark:text-primary-dark/75">
-                                                    {feature.name}
-                                                </strong>
-                                            </Tooltip>
-                                        ) : (
-                                            <strong className="text-primary/75 dark:text-primary-dark/75">
-                                                {feature.name}
-                                            </strong>
-                                        )}
-                                    </div>
-                                    {platformAndSupportProduct?.plans
-                                        ?.filter((plan: BillingV2PlanType) => plan.name !== 'Teams') // This is a temporary addition until the teams addon is shipped and the teams plan is removed
-                                        ?.map((plan: BillingV2PlanType) => {
-                                            const planFeature = plan?.features?.find((f) => f.key === feature.key)
-                                            return (
-                                                <div
-                                                    className="col-span-4 px-3 py-2 text-sm"
-                                                    key={`${plan.key}-${feature.key}`}
-                                                >
-                                                    {planFeature ? (
-                                                        <div className="flex gap-x-2">
-                                                            {planFeature.note ?? (
-                                                                <IconCheck className="w-5 h-5 text-green" />
-                                                            )}
-                                                            {planFeature.limit && (
-                                                                <span className="opacity-75">
-                                                                    <>
-                                                                        {planFeature.limit} {planFeature.unit}
-                                                                    </>
-                                                                </span>
-                                                            )}
-                                                        </div>
-                                                    ) : (
-                                                        <></>
-                                                    )}
-                                                </div>
-                                            )
-                                        })}
-                                </>
-                            ))}
-                    </div>
-                </div>
-            </section>
-        </>
-    )
-}
-
-const tiers = [
-    {
-        icon: <IconGraph className="text-blue" />,
-        title: 'Analytics',
-        startsAt: <><strong className="text-sm">$0.00031</strong><span className="opacity-60 text-[13px]">/event (or cheaper)</span></>,
-        children: (
-            <>
-                content
-            </>
-        )
-    },
-    {
-        icon: <IconRewindPlay className="text-yellow" />,
-        title: 'Session replay',
-        startsAt: <><strong className="text-sm">$0.0050</strong><span className="opacity-60 text-[13px]">/recording</span></>,
-        children: (
-            <>
-                content
-            </>
-        )
-    },
-    {
-        icon: <IconToggle className="text-seagreen" />,
-        title: 'Feature flags',
-        startsAt: <><strong className="text-sm">$0.0001</strong><span className="opacity-60 text-[13px]">/recording</span></>,
-        children: (
-            <>
-                content
-            </>
-        )
-    },
-    {
-        icon: <IconFlask className="text-purple" />,
-        title: 'A/B testing',
-        children: (
-            <>
-                content
-            </>
-        )
-    },
-    {
-        icon: <IconMessage className="text-salmon" />,
-        title: 'Surveys',
-        startsAt: <><strong className="text-sm">$0.2000</strong><span className="opacity-60 text-[13px]">/response</span></>,
-        children: (
-            <>
-                content
-            </>
-        )
-    }
-
-]
-
-interface PlanData {
-    title: string
-    price: string
-    priceSubtitle?: string | JSX.Element
-    features: React.ReactNode[]
-    CTAText?: string
-    CTALink?: string
-}
-
-const planSummary = [
-    {
-        title: 'Totally free',
-        price: 'Free',
-        priceSubtitle: '- no credit card required',
-        features: [
-            'Generous free monthly tier',
-            'Basic product features',
-            '1 project',
-            '1 year data retention',
-            'Community support',
-        ],
-    },
-    {
-        title: 'Ridiculously cheap',
-        price: '$0',
-        features: [
-            'Generous free tier on all products',
-            'Advanced product features',
-            '6 projects',
-            '7 year data retention',
-            'Email support',
-            'Pay only for what you use',
-            <>
-                <span className="opacity-60 text-sm">* Included with any product subscription</span>
-            </>,
-        ],
-    },
-    {
-        title: 'Enterprise',
-        price: 'Custom pricing',
-        priceSubtitle: 'w/ fixed terms',
-        features: [
-            'Unlimited everything',
-            'SAML SSO',
-            'Custom MSA',
-            'Dedicated support',
-            'Personalized onboarding & training',
-            'Advanced permissions & audit logs',
-        ],
-        CTAText: 'Get in touch',
-        CTALink: '/talk-to-a-human',
-    },
-]
-
-const Plan: React.FC<{ planData: PlanData }> = ({ planData }) => (
-    <div>
-        <h4 className="text-lg mb-2">{planData.title}</h4>
-        <div className="flex flex-col h-full border border-light dark:border-dark bg-white dark:bg-accent-dark rounded">
-            <div className="flex flex-col h-full gap-4 pt-3 px-4 xl:px-4 pb-6">
-                <div>
-                    <h4 className="inline text-lg">
-                        {planData.price != 'Free' && planData.price != 'Custom pricing' && (
-                            <span className="text-sm opacity-60 font-normal">Starts at</span>
-                        )}{' '}
-                        {planData.price}
-                        {planData.price != 'Free' && planData.price != 'Custom pricing' && (
-                            <span className="text-sm opacity-60 font-normal">/mo</span>
-                        )}
-                    </h4>
-                    &nbsp;
-                    <p className="inline opacity-75 text-sm">{planData.priceSubtitle}</p>
-                </div>
-                <ul className="p-0 list-none flex flex-col gap-2 [&_li]:text-sm xl:[&_li]:text-[15px]">
-                    {planData.features.map((feature, index) => (
-                        <li key={index}>{feature}</li>
-                    ))}
-                </ul>
-                <div className="mt-auto">
-                    <PlanCTA
-                        type={planData.title === 'Ridiculously cheap' ? 'primary' : 'secondary'}
-                        ctaText={planData.CTAText}
-                        ctaLink={planData.CTALink}
-                    />
-                </div>
-            </div>
-        </div>
-    </div>
-)
+import { PlanColumns } from './Test/PlanColumns';
+import { FreePlanContent, PaidPlanContent } from './Test/PlanContent'
+import { Accordion, tiers } from './Test/PricingAccordion';
 
 const SectionLayout = ({ id = '', children }) => (
     <section id={id} className={`${section} mb-12 mt-8 md:px-4`}>
@@ -1087,209 +802,9 @@ const PricingExperiment = ({
         )
     }
 
-    const FreePlanContent = () => {
-        return (
-            <>
-                <div>
-                    <h4 className="text-xl mb-0">Free</h4>
-                    <p className="text-sm opacity-75">No credit card required</p>
 
-                    <ul className="list-none p-0 mb-4 space-y-1">
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            <span>Generous <button className="text-red dark:text-yellow font-semibold">monthly free tier</button></span>
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            Community support
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            1 project, 1-year data retention
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            Unlimited team members
-                        </li>
-                    </ul>
-                </div>
-                <div className="mt-auto">
-                    <div className="mb-4">
-                        <label className="block opacity-75 text-[15px] mb-1">Cloud region</label>
-                        <div className="flex gap-2">
-                            <button className="flex-1 flex flex-col items-center py-1.5 px-2 text-sm rounded border border-yellow bg-yellow/25 dark:bg-white/5 font-bold">
-                                US (Virginia)
-                            </button>
-                            <button className="flex-1 flex flex-col items-center py-1.5 px-2 text-sm rounded border border-light hover:border-dark/50 dark:border-dark dark:hover:border-light/50 bg-transparent">
-                                EU (Frankfurt)
-                            </button>
-                        </div>
-                    </div>
-                    <PlanCTA intent="free" size="lg" width="full" />
-                    <p className="text-sm text-center mt-4 mb-0 opacity-75"><strong>128 companies</strong> signed up today</p>
-                </div>
-            </>
-        )
-    }
 
-    const PaidPlanContent = () => {
-        return (
-            <>
-                <div>
-                    <span className="text-70 text-[15px]">From</span>
-                    <div className="flex items-baseline">
-                        <h4 className="text-xl mb-0">$0</h4>
-                        <span className="opacity-70 text-sm">/mo</span>
-                    </div>
-                    <p className="text-sm mt-2">
-                        <button className="text-red dark:text-yellow font-semibold text-[15px]">Estimate your price</button>
-                    </p>
 
-                    <ul className="list-none p-0 mb-4 space-y-1">
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            <span>Generous <button className="text-red dark:text-yellow font-semibold">monthly free tier</button></span>
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            <span>
-                                Then, <strong className="bg-yellow/50 dark:bg-white/50 p-0.5 font-semibold">usage-based pricing</strong>
-                            </span>
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            <span>
-                                <strong className="bg-yellow/50 dark:bg-white/50 p-0.5 font-semibold">Email</strong> support
-                            </span>
-                        </li>
-                        <li className="flex items-start gap-2 text-sm">
-                            <IconCheck className="text-green inline-block size-5" />
-                            Unlimited team members
-                        </li>
-                    </ul>
-                </div>
-                <div className="mt-auto">
-                    <div className="mb-4">
-                        <label className="block opacity-75 text-[15px] mb-1">Cloud region</label>
-                        <div className="flex gap-2">
-                            <button className="flex-1 flex flex-col items-center py-1.5 px-2 text-sm rounded border border-yellow bg-yellow/25 dark:bg-white/5 font-bold">
-                                US (Virginia)
-                            </button>
-                            <button className="flex-1 flex flex-col items-center py-1.5 px-2 text-sm rounded border border-light hover:border-dark/50 dark:border-dark dark:hover:border-light/50 bg-transparent">
-                                EU (Frankfurt)
-                            </button>
-                        </div>
-                    </div>
-                    <PlanCTA intent="free" size="lg" width="full" />
-                    <p className="text-sm text-center mt-4 mb-0 opacity-75"><strong>128 companies</strong> signed up today</p>
-                </div>
-            </>
-        )
-    }
-
-    const AccordionItem = ({
-        icon,
-        title,
-        startsAt = '',
-        description,
-        children,
-        isOpen,
-        onClick,
-        onAnimationComplete,
-    }) => {
-        const contentRef = useRef(null)
-
-        return (
-            <li
-                className={`border-t relative ${isOpen
-                    ? 'active border-transparent bg-white dark:bg-accent-dark rounded shadow-lg z-10 overflow-hidden'
-                    : 'inactive border-light dark:border-dark first:border-transparent'
-                    }`}
-            >
-                <button
-                    onClick={onClick}
-                    className={`text-left pl-3 pr-4 cursor-pointer w-full flex justify-between items-center transition-all rounded relative ${isOpen
-                        ? 'pt-4 pb-2 z-20'
-                        : 'text-primary/90 hover:text-primary/100 dark:text-primary-dark/90 dark:hover:text-primary-dark/100 py-2 hover:bg-accent/80 dark:hover:bg-accent/5 hover:scale-[1.0025] hover:top-[-.5px] active:scale-[.9999] active:top-[3px]'
-                        }`}
-                >
-                    <div className="grid grid-cols-12 w-full gap-1 items-center">
-                        <div className="col-span-5">
-                            <div className="flex gap-1 items-center">
-                                <div className={isOpen ? 'size-6' : 'size-5'}>{icon}</div>
-                                <span
-                                    className={`transition-all leading-tight font-bold ${isOpen ? 'text-base md:text-[17px]' : 'text-[15px] md:text-base'
-                                        }`}
-                                >
-                                    {title}
-                                </span>
-                            </div>
-                        </div>
-                        <div className="col-span-6">
-                            <span>
-                                {startsAt}
-                            </span>
-                        </div>
-                        <span>
-                            {isOpen ? (
-                                <IconMinus className="size-4 inline-block transform rotate-180" />
-                            ) : (
-                                <IconPlus className="size-4 inline-block transform rotate-0" />
-                            )}
-                        </span>
-                    </div>
-
-                </button>
-                <motion.div
-                    onAnimationComplete={onAnimationComplete}
-                    ref={contentRef}
-                    animate={{ height: isOpen ? 'auto' : 0, transition: { duration: 0.3, type: 'tween' } }}
-                    className={isOpen ? '' : 'overflow-hidden'}
-                >
-                    <div className="px-4 pb-4">
-                        {children}
-                    </div>
-                </motion.div>
-            </li>
-        )
-    }
-
-    const Accordion = ({ items }) => {
-        const ref = useRef<HTMLOListElement>(null)
-        const [openIndex, setOpenIndex] = useState(null)
-
-        const scrollToIndex = (index) => {
-            if (ref.current && window.innerWidth <= 639) {
-                const element = ref.current.children[index]
-                const y = element.getBoundingClientRect().top + window.scrollY - 56
-                window.scrollTo({ top: y, behavior: 'smooth' })
-            }
-        }
-
-        return (
-            <ol ref={ref} className="space-y-px p-0 list-none">
-                {items.map((item, index) => (
-                    <AccordionItem
-                        onAnimationComplete={({ height }) => {
-                            if (height === 'auto') {
-                                scrollToIndex(index)
-                            }
-                        }}
-                        key={index}
-                        icon={item.icon}
-                        title={item.title}
-                        startsAt={item.startsAt}
-                        isOpen={openIndex === index}
-                        onClick={() => {
-                            setOpenIndex(openIndex === index ? null : index)
-                        }}
-                    >
-                        {item.children}
-                    </AccordionItem>
-                ))}
-            </ol>
-        )
-    }
